[layouts/type-stage] more renaming typespec-ramp -> type-stage.
