+++ lib/assets/type-stage-initial-state.json
