+++ lib/js/components/type-spec-models.mjs
@@ -292,7 +292,7 @@ export const StylePatchTypeModel = _AbstractGenericModel.createClass('StylePatch
         // into a
 
         // FIXME: could we define a list of keys into
-           // the dependency "stylePatches" of the parent TypeSpecRampModel // a StylePatchesMapModel
+           // the dependency "stylePatches" of the parent TypeStageModel // a StylePatchesMapModel
            // and SET_NULL + ALLOW_NULL seems not possible to me now, but I can look it up
            // we could read the styles directly, but I don't see how dependencies are
            // resolved smartly on metamorphose.
@@ -358,7 +358,7 @@ export const StylePatchTypeModel = _AbstractGenericModel.createClass('StylePatch
         // To be able to reference the available patches, we should include
         // a dependency to the StylePatchesMapModel keys within that are
         // where the keys can be identifiers like "italic", "quote", "strong", "code"
-        // and the values are keys in TypeSpecRampModel/stylePatches
+        // and the values are keys in TypeStageModel/stylePatches
         // if the key does not exist it's a null style, no need
         // to set null, the style just doesn't do anything. That way editing
         // is simple as well. Like FOREIGN_KEY_NO_ACTION
